+++ src/main/java/storage/IndexedPage.java
@@ -3,5 +3,5 @@ package storage;
 public record IndexedPage(Endpoint left, Endpoint right, int pageNumber) {
   public sealed interface Endpoint permits Unbounded, Bounded {}
   public record Unbounded() implements Endpoint {}
-  public record Bounded(int endpoint) implements Endpoint {}
+  public record Bounded(long endpoint) implements Endpoint {}
 }
