+++ src/main/java/storage/Value.java
@@ -2,12 +2,12 @@ package storage;
 
 public sealed interface Value {
   record NullValue() implements Value {}
-  record IntValue(int value) implements Value {}
+  record IntValue(long value) implements Value {}
   record BlobValue(byte[] blob) implements Value {}
   record StringValue(String data) implements Value {}
 
   default String getString() {return ((StringValue) this).data;}
-  default int getInt() {return ((IntValue) this).value;}
+  default long getInt() {return ((IntValue) this).value;}
 
   default String display() {
     return switch (this) {
